refactor:Use custom hooks to isolate data-fetching logic
